requirements.txt
diffusers>=0.30.0
transformers>=4.40.0
dataclasses-json==0.6.7
uvicorn>=0.23.0
fastapi==0.108.0
strawberry-graphql==0.235.1
numpy>=1.23.5
